Update all model classes to Electron API 39.2

+++ src/ElectronNET.API/API/Entities/UpdateInfo.cs
@@ -3,36 +3,37 @@
     /// <summary>
     /// 
     /// </summary>
+    /// <remarks>Up-to-date with electron-updater 6.7.2</remarks>
     public class UpdateInfo
     {
         /// <summary>
-        /// The version.
+        /// Gets or sets the version.
         /// </summary>
         public string Version { get; set; }
 
         /// <summary>
-        /// 
+        /// Gets or sets the files included in this update.
         /// </summary>
         public UpdateFileInfo[] Files { get; set; } = new UpdateFileInfo[0];
 
         /// <summary>
-        /// The release name.
+        /// Gets or sets the release name.
         /// </summary>
         public string ReleaseName { get; set; }
 
         /// <summary>
-        /// The release notes.
+        /// Gets or sets the release notes.
         /// </summary>
         public ReleaseNoteInfo[] ReleaseNotes { get; set; } = new ReleaseNoteInfo[0];
 
         /// <summary>
-        /// 
+        /// Gets or sets the release date.
         /// </summary>
         public string ReleaseDate { get; set; }
 
         /// <summary>
-        /// The staged rollout percentage, 0-100.
+        /// Gets or sets the staged rollout percentage, 0-100.
         /// </summary>
-        public int StagingPercentage { get; set; }
+        public double StagingPercentage { get; set; }
     }
 }
